requirements.txt
pandas
streamlit
numpy
watchdog
openpyxl
plotly
python-docx
tqdm
beautifulsoup4
langdetect
kaleido
autopep8
autoflake
isort
black
xlsxwriter
fastapi
uvicorn
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib
aiofiles
